# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PDD Petty Cash API
  description: |
    REST API for the PDD Petty Cash (Travel Expense Management) app.
    See CLAUDE.md §9 for conventions: camelCase JSON, RFC 7807 errors, cursor pagination,
    Idempotency-Key required on POST /expenses, /transfers, /allocations.
  version: 0.1.0
servers:
  - url: https://api.pettycash.pdd.gov.ae/api/v1
    description: Production (Moro Hub) — placeholder
  - url: http://localhost:8080/api/v1
    description: Local Spring Boot
...
security:
  - bearerAuth: []
paths:
...
  /trips/{tripId}/expenses:
    get:
      tags: [expenses]
      summary: List expenses for a trip, cursor-paginated
      description: |
        Cursor pagination per CLAUDE.md §9. Rows ordered by
        `(occurredAt DESC, id DESC)`. To fetch subsequent pages, echo the
        `nextCursor` from the previous response back as the `cursor` query
        param. Cursor is opaque base64 — clients must not parse it.
        Returns `400 INVALID_CURSOR` on a malformed cursor.

        `limit` is clamped to [1, 100]; default 20.
      parameters:
        - { $ref: '#/components/parameters/TripIdPath' }
        - in: query
          name: userId
          schema: { type: string }
        - in: query
          name: scope
          schema: { type: string, enum: [mine, all], default: all }
        - in: query
          name: cursor
          description: Opaque cursor from a previous page's `nextCursor`. Omit on first page.
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
        - in: query
          name: categoryCode
          schema: { type: string }
        - in: query
          name: sourceId
          schema: { type: string }
        - in: query
          name: memberId
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExpensePage' }
    post:
      tags: [expenses]
      summary: Create an expense (idempotent, client-supplied UUID)
      parameters:
        - { $ref: '#/components/parameters/TripIdPath' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateExpenseRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Expense' }
...
  /expenses/{expenseId}:
    get:
      tags: [expenses]
      parameters: [{ $ref: '#/components/parameters/ExpenseIdPath' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Expense' }
    patch:
      tags: [expenses]
      parameters: [{ $ref: '#/components/parameters/ExpenseIdPath' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExpensePatch' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Expense' }
  /expenses/{expenseId}/source:
    patch:
      tags: [expenses]
      summary: Reassign source (audited event per CLAUDE.md §5)
      parameters: [{ $ref: '#/components/parameters/ExpenseIdPath' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sourceId]
              properties:
                sourceId: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Expense' }
...
  /categories:
    get:
      tags: [categories]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ExpenseCategory' }
    post:
      tags: [categories]
      summary: Add a new category (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCategoryRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExpenseCategory' }
...
  /chat/threads:
    get:
      tags: [chat]
      parameters:
        - in: query
          name: tripId
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ChatThread' }
  /chat/threads/{threadId}/messages:
    get:
      tags: [chat]
      parameters:
        - in: path
          name: threadId
          required: true
          schema: { type: string }
        - in: query
          name: cursor
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50, maximum: 200 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ChatMessage' }
    post:
      tags: [chat]
      parameters:
        - in: path
          name: threadId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChatMessage' }
  /chat/threads/{threadId}/poll:
    get:
      tags: [chat]
      summary: Long-poll for new messages in a chat thread
      description: |
        Returns immediately if any messages exist with `sentAt > since` for
        this thread. Otherwise hangs for up to `timeoutSeconds` (default 25,
        max 30) and resolves when a peer posts, or with an empty `items`
        array on timeout. Thread membership is enforced — non-participants
        get `403 FORBIDDEN`.

        Same single-instance caveat as `/notifications/poll`.
      parameters:
        - in: path
          name: threadId
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: since
          schema: { type: string, format: date-time }
        - in: query
          name: timeoutSeconds
          schema: { type: integer, default: 25, minimum: 1, maximum: 30 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items, serverNow]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ChatMessage' }
                  serverNow:
                    type: string
                    format: date-time
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TripIdPath:
      in: path
      name: tripId
      required: true
      schema: { type: string }
    ExpenseIdPath:
      in: path
      name: expenseId
      required: true
      schema: { type: string }
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema: { type: string, format: uuid }
  responses:
    Problem:
      description: RFC 7807 problem details with a stable `code` field per CLAUDE.md §9.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            idempotencyKeyRequired:
              summary: IDEMPOTENCY_KEY_REQUIRED (400)
              value:
                type: https://api.pettycash.pdd.gov.ae/problems/idempotency_key_required
                title: Bad Request
                status: 400
                detail: "Header 'Idempotency-Key' is required on this endpoint"
                instance: /api/v1/trips/abc/expenses
                code: IDEMPOTENCY_KEY_REQUIRED
            idempotencyKeyConflict:
              summary: IDEMPOTENCY_KEY_CONFLICT (409)
              value:
                type: https://api.pettycash.pdd.gov.ae/problems/idempotency_key_conflict
                title: Conflict
                status: 409
                detail: "Idempotency-Key reused with a different request body"
                instance: /api/v1/trips/abc/expenses
                code: IDEMPOTENCY_KEY_CONFLICT
            reportDeferred:
              summary: REPORT_DEFERRED (501)
              value:
                type: https://api.pettycash.pdd.gov.ae/problems/report_deferred
                title: Not Implemented
                status: 501
                detail: "Server-side report generation pending — see CLAUDE.md §10."
                instance: /api/v1/reports/trip/abc
                code: REPORT_DEFERRED
  schemas:
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: integer
          format: int64
          description: Minor units (e.g. halalas for SAR, fils for AED)
        currency:
          type: string
          example: SAR
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string }
...
    ExpenseCategory:
      type: object
      properties:
        code: { type: string }
        nameEn: { type: string }
        nameAr: { type: string }
        iconKey: { type: string }
        isActive: { type: boolean }
    CreateCategoryRequest:
      type: object
      required: [code, nameEn, nameAr, iconKey]
      properties:
        code: { type: string }
        nameEn: { type: string }
        nameAr: { type: string }
        iconKey: { type: string }
    Expense:
      type: object
      properties:
        id: { type: string }
        tripId: { type: string }
        userId: { type: string }
        sourceId: { type: string }
        categoryCode: { type: string }
        amount: { $ref: '#/components/schemas/Money' }
        quantity: { type: integer, default: 1 }
        details: { type: string }
        vendor:
          type: string
          maxLength: 200
          description: Optional vendor / merchant name (free-text, ≤200 chars).
        occurredAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        receiptObjectKey: { type: string }
    ExpensePage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Expense' }
        nextCursor: { type: string }
    CreateExpenseRequest:
      type: object
      required: [id, sourceId, categoryCode, amount, occurredAt]
      properties:
        id:
          type: string
          description: Client-supplied UUID; becomes the server-side Expense.id.
        sourceId: { type: string }
        categoryCode: { type: string }
        amount: { $ref: '#/components/schemas/Money' }
        quantity: { type: integer, default: 1 }
        details: { type: string }
        vendor:
          type: string
          maxLength: 200
        occurredAt: { type: string, format: date-time }
        receiptObjectKey: { type: string }
    ExpensePatch:
      type: object
      properties:
        sourceId: { type: string }
        categoryCode: { type: string }
        amount: { $ref: '#/components/schemas/Money' }
        details: { type: string }
        vendor:
          type: string
          maxLength: 200
        occurredAt: { type: string, format: date-time }
...
    ChatThread:
      type: object
      properties:
        id: { type: string }
        tripId: { type: string }
        title: { type: string }
        titleAr: { type: string }
        participantIds:
          type: array
          items: { type: string }
        unreadCount: { type: integer }
        lastMessagePreview: { type: string }
        lastMessageAt: { type: string, format: date-time }
    ChatMessage:
      type: object
      properties:
        id: { type: string }
        threadId: { type: string }
        senderId: { type: string }
        body: { type: string }
        sentAt: { type: string, format: date-time }
        deliveredAt: { type: string, format: date-time }
        readAt: { type: string, format: date-time }
...